New KGVC program: no branch parameter
KGVC will now only accept 3 parameters. Getting rid of the branch parameter. The [project] parameter is now refactored too as: workspace [desPath]. The system has a preference in which allows the number of certain depths of accessing directories, in which the program splits up the given user workspace [desPath] by each os file separator until the final path which is the workspace.